requirements.txt
asgiref==3.8.1
Django==5.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.1
pillow==10.4.0
PyJWT==2.9.0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
